requirements.txt
groq==0.31.0
httpx==0.27.0
python-dotenv==1.0.1
